opennaskのフロントエンドコードを分割して見通しをよくする
